Redirect back to account settings when exiting the Theme Editor
refs CNVS-22055

This change replaces the 'X' close button with a 'Cancel' button.
Also now, whenever you exit out of the Theme Editor, via either
'Apply' or 'Cancel' you will be redirected back to the account
settings page instead of the preview iframe location.

These changes should make things work better once CNVS-2196 is
implemented.

This change also cleans up the Theme Editor header a bit, adding
the Canvas logo, and adjusting some padding/margins.

Test plan:
1. Open Theme Editor and click 'Cancel'. You should be redirected back
account settings page and the appropriate message should display.
2. Open Theme Editor, make some changes, click 'Preview', click 'Cancel'.
You should be redirected to the account settings page, the theme changes
should be reset/removed and the appropriate message should display.
3. Open Theme Editor, make some changes, click 'Preview', click 'Apply'.
You should be redirected to the account settings page and the appropriate
message should display.

Reviewed-on: https://gerrit.instructure.com/58913
Tested-by: Jenkins
Product-Review: Colleen Palmer <colleen@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>

+++ app/controllers/brand_configs_controller.rb
@@ -72,16 +72,17 @@ class BrandConfigsController < ApplicationController
     @domain_root_account.brand_config = new_md5 && BrandConfig.find(new_md5)
     @domain_root_account.save!
     BrandConfig.destroy_if_unused(old_md5)
-    redirect_to :back, notice: t('Success! All users on this domain will now see this branding.')
+    redirect_to account_path(@domain_root_account), notice: t('Success! All users on this domain will now see this branding.')
   end
 
   # When you close the theme editor, it will send a DELETE to this action to
   # clear out the session brand_config that you were prevewing.
   def destroy
-    session.delete(:brand_config_md5)
-    BrandConfig.destroy_if_unused(session.delete(:brand_config_md5))
-    flash[:notice] = t('Theme editor changes have been cancelled.')
-    render json: {success: true}
+    if session.delete(:brand_config_md5).presence
+      session.delete(:brand_config_md5)
+      BrandConfig.destroy_if_unused(session.delete(:brand_config_md5))
+    end
+    redirect_to account_path(@domain_root_account), notice: t('Theme editor changes have been cancelled.')
   end
 
   protected
